requirements.txt
python-telegram-bot==21.11.1
python-dotenv==1.0.1
openpyxl==3.1.5
reportlab==4.4.1
